wip: product service

+++ product/src/main/java/de/hska/iwi/EShop/product/persistence/ProductRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.repository.ListCrudRepository;
 
 public interface ProductRepository extends ListCrudRepository<Product, Integer> {
     @Query(value = "SELECT * FROM product WHERE price >= ?1 AND price <= ?2 WHERE name LIKE '%?3%'")
-    List<Product> getFilteredProducts(Integer minPrice, Integer maxPrice, String searchText);
+    List<Product> getFilteredProducts(Double minPrice, Double maxPrice, String searchText);
 }
